# backend/swagger.yaml
openapi: 3.0.3
info:
  title: ChatGPT Clone API
  description: Backend API for ChatGPT Clone with AI integration, file uploads, and real-time chat
  version: 1.0.0
  contact:
    name: Galaxy Assignment
    email: support@galaxy-assignment.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "64f8d123abc456def789"
        clerkId:
          type: string
          example: "user_2abc123def456"
        email:
          type: string
          format: email
          example: "user@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        avatar:
          type: string
          format: uri
          example: "https://images.clerk.dev/avatar.jpg"
        preferences:
          type: object
          properties:
            model:
              type: string
              example: "gpt-3.5-turbo"
            theme:
              type: string
              example: "light"
        subscription:
          type: string
          enum: [free, pro, enterprise]
          example: "free"
        usage:
          type: object
          properties:
            messagesCount:
              type: integer
              example: 150
            filesUploaded:
              type: integer
              example: 5
            lastActive:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid input provided"
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /api/users/profile:
    get:
      summary: Get User Profile
      description: Retrieve current user's profile information
      tags:
        - Users
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update User Profile
      description: Update user profile information
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preferences:
                  type: object
                  properties:
                    model:
                      type: string
                      example: "gpt-4"
                    theme:
                      type: string
                      example: "dark"
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  preferences:
                    type: object
                  updatedAt:
                    type: string
                    format: date-time
...